video-streaming-service/requirements.txt
Flask>=2.2.2
mysql-connector-python==8.0.23
requests==2.25.1